Fix the WebSocketException exception
When trying to recreate another thread to run ws.run_forever, I got
WebSocketException: socket is already opened. It is because when
on_close is been called, the ws.sock is is not None. It will be set to
None when on_close is returned. But the ws.run_forever checks the
ws.sock at the beginning of the method. And I cannot join the websocket
event loop thread in on_close, because on_close is  running inside the
event loop thread, it cannot join itself. So I create a new thread to
join the previous event loop thread, and then run websocket event loop.